+++ CHANGELOG.md
@@ -6,8 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2022-12-25
+### Added
+- Link to this changelog (https://github.com/Gimanua/alko-logg/blob/main/CHANGELOG.md) on the version text.
+- Support for local storage, automatically saves and loads the list of drinks to and from the local storage.
+- Editing the count to 0 of a drinks removes it from the list.
+- Button for clearing the whole list of drinks
+
+### Changed
+- 0 is now a valid value for volume and alchol content
+
 ## [1.1.0] - 2022-10-14
-###
+### Added
 - Ability to click on table cells to edit the values and recalculate the total shots value.
 
 ## [1.0.0] - 2022-08-04